Create API tests - Add test to `GET /v1/problems` - Add `supertest` to `package.json` - Export `app` from `app.ts` and run server in `index.ts`
